# Conflicts:
#	src/main/java/prati/projeto/redeSocial/rest/controller/AvaliacaoController.java
#	src/main/java/prati/projeto/redeSocial/rest/controller/ComentarioRespostaController.java
#	src/main/java/prati/projeto/redeSocial/rest/controller/RelacionamentoSeguidoresController.java
#	src/main/java/prati/projeto/redeSocial/rest/dto/AvaliacaoDTO.java
#	src/main/java/prati/projeto/redeSocial/rest/dto/ResenhaDTO.java
#	src/main/java/prati/projeto/redeSocial/rest/dto/ResenhaViewDTO.java

+++ src/main/java/prati/projeto/redeSocial/modal/entity/Resenha.java
@@ -47,7 +47,7 @@ public class Resenha {
     private List<Avaliacao> avaliacoes = new ArrayList<>();
 
     @Column(name = "media_resenha")
-    private Integer media = 0;
+    private Double media = 0.0;
 
     @Column(name = " possui_spoiler")
     private boolean spoiler;
